selection sort

+++ 2-selection_sort.c
@@ -1,5 +1,4 @@
 #include "sort.h"
-#include <stddef.h>
 
 /**
  * selection_sort - sort an array of integers in ascending order
@@ -12,18 +11,18 @@
 
 void selection_sort(int *array, size_t size)
 {
-	size_t i, j, pos;
+	int i, j, pos;
 	int min, temp, flag;
 
 	if (size < 2)
 		return;
-	for (i = 0; i < size; i++)
+	for (i = 0; (size_t)i < size; i++)
 	{
 		flag = 0;
 		min = array[i];
-		if (i == size - 1)
+		if ((size_t)i == size - 1)
 			break;
-		for (j = i; j < size; j++)
+		for (j = i; (size_t)j < size; j++)
 		{
 			if (min > array[j])
 			{
